+++ templates/single-fitting.php
@@ -44,6 +44,7 @@ get_header();
                 ?>
             </div> <!-- /.content -->
         </div> <!-- /.col-lg-9 /.col-md-9 /.col-sm-9 /.col-9 -->
+
         <?php
         if(PluginHelper::getInstance()->hasSidebar('sidebar-fitting-manager')) {
             ?>